use floats where needed in custom actions (#2371)

+++ src/queries/participants/cancel-individual-from-course.ts
@@ -5,7 +5,7 @@ export const CANCEL_INDIVIDUAL_FROM_COURSE_MUTATION = gql`
     $courseId: Int!
     $profileId: uuid!
     $reason: String!
-    $fee: Int!
+    $fee: Float!
   ) {
     cancelIndividualFromCourse(
       courseId: $courseId
